test: refactor timout test, split basic keyless into a separate test
